Everything in Vane/ has been moved into Vane/Vane (for better, less ambiguous #include directives).
