+++ Sandbox/main.cpp
@@ -1,9 +1,9 @@
 #include <iostream>
 
-#include <Core/Application.hpp>
-#include <Core/Logger.hpp>
-#include <Platform/Platform.hpp>
-#include <EntryPoint.hpp>
+#include <Vane/Core/Application.hpp>
+#include <Vane/Core/Logger.hpp>
+#include <Vane/Platform/Platform.hpp>
+#include <Vane/EntryPoint.hpp>
 
 class Game : public Vane::Application {
 public:
