--- Vane/Core/Logger.cpp
@@ -1,45 +1,0 @@
-//////////////////////////////////////////////////////////////////////////////////////////
-// Copyright wowvain-dev | Bogdan Stanciu (c) 2024.
-//
-// MIT License
-// Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the “Software”), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-//////////////////////////////////////////////////////////////////////////////////////////
-
-#include "Logger.hpp"
-
-#include <Core/Asserts.hpp>
-
-namespace Vane {
-void reportAssertionFailure(const char* expression, const char* message, const char* file, i32 line) {
-    Logger::log_output(LOG_LEVEL::V_FATAL, "Assertion Failure: {}, message: '{}', in file: {}, line: {}\n",
-                       expression,
-                       message,
-                       file,
-                       line
-    );
-}
-
-void reportAssertionFailure(
-    std::string expression,
-    std::string message,
-    std::string file,
-    i32 line
-) {
-    Logger::log_output(LOG_LEVEL::V_FATAL, "Assertion Failure: {}, message: '{}', in file: {}, line: {}\n",
-                       expression,
-                       message,
-                       file,
-                       line
-    );
-}
-
-bool Logger::initializeLogging() {
-    return true;
-}
-
-void Logger::shutdownLogging() {}
-} // Vane
