+++ Vane/Vane/Core/Memory/ObjectHandle.cpp
