+++ Vane/Vane/Platform/Platform.hpp
@@ -25,7 +25,7 @@
 #include <iostream>
 #include <string>
 
-#include <Core/Defines.hpp>
+#include <Vane/Core/Defines.hpp>
 
 #ifdef VPLATFORM_LINUX
