+++ Vane/Vane/Platform/Platform_Win32.cpp
@@ -2,7 +2,7 @@
 
 #ifdef VPLATFORM_WINDOWS
 
-#include <Core/Logger.hpp>
+#include <Vane/Core/Logger.hpp>
 
 using namespace Vane;
